:recycle: version3(pulse) Refactoring code.
