+++ pkg/subscription/subscription.go
@@ -8,8 +8,8 @@ import (
 	"github.com/silverswords/pulse/pkg/driver"
 	"github.com/silverswords/pulse/pkg/logger"
 	"github.com/silverswords/pulse/pkg/message"
-	"github.com/silverswords/pulse/pkg/retry"
-	"github.com/silverswords/pulse/pkg/scheduler"
+	"github.com/silverswords/pulse/pkg/protocol/retry"
+	"github.com/silverswords/pulse/pkg/protocol/scheduler"
 	"github.com/silverswords/pulse/pkg/topic"
 	"github.com/valyala/fasthttp"
 	"sync"
