+++ pkg/topic/topic.go
@@ -6,12 +6,12 @@ import (
 	"errors"
 	"fmt"
 	jsoniter "github.com/json-iterator/go"
-	"github.com/silverswords/pulse/pkg/deadpolicy"
 	"github.com/silverswords/pulse/pkg/driver"
 	"github.com/silverswords/pulse/pkg/logger"
 	"github.com/silverswords/pulse/pkg/message"
-	"github.com/silverswords/pulse/pkg/retry"
-	"github.com/silverswords/pulse/pkg/scheduler"
+	"github.com/silverswords/pulse/pkg/protocol/deadpolicy"
+	"github.com/silverswords/pulse/pkg/protocol/retry"
+	"github.com/silverswords/pulse/pkg/protocol/scheduler"
 	"golang.org/x/sync/errgroup"
 
 	//"go.opencensus.io/stats"
